20160510 18:36

+++ proj.android/copy_assert.sh
@@ -5,12 +5,6 @@ echo "APP_ROOT = $APP_ROOT"
 echo "APP_ANDROID_ROOT = $APP_ANDROID_ROOT"
 # make sure assets is exist
 
-if [ ! $3 ]
-	then
-	echo "第三个参数渠道号"
-	exit 2
-fi
-
 function copyAssert()
 {
 	echo "copy $1 to $2"
@@ -24,10 +18,9 @@ function copyAssert()
 		cp "$file" "$APP_ANDROID_ROOT"/$2
 	fi
 	done
-	# chmod 777 -R "$APP_ANDROID_ROOT"/$2
+	chmod 777 -R "$APP_ANDROID_ROOT"/$2
 	echo "end copy $1"
 }
-
 function encodeImageAndCopy()
 {
  echo "encodeImage $1 $2 "
@@ -47,15 +40,12 @@ EOF
 fi
 cd $APP_ANDROID_ROOT
 }
-
 echo "clean dirty file"
 if [ -d "$APP_ANDROID_ROOT"/assets ]; then
     rm -rf "$APP_ANDROID_ROOT"/assets
 fi
 
 mkdir "$APP_ANDROID_ROOT"/assets
-
-# sh ./copy_scripts.sh
 
 cd ../scripts
 rm -rf ../build
@@ -70,19 +60,59 @@ cd ../proj.android
 echo "end build"
 
 copyAssert build assets
+copyAssert android_sound assets
+if [ $1 = 0 ]
+	then
+	copyAssert res assets
+	copyAssert unPackagedRes/music assets
+	copyAssert unPackagedRes/res assets
+	
+	echo "scriptsVersion = 999999" >$APP_ANDROID_ROOT/assets/config.lua
+	 echo "module('release', package.seeall) release = false" > $APP_ANDROID_ROOT/assets/release.lua
+fi
+if [ $1 = 1 ]
+	then
+	copyAssert res assets
+#	//cd ..
+	rm -rf "$APP_ROOT"/encodeTmp
+	mkdir "$APP_ROOT"/encodeTmp
+	mkdir "$APP_ROOT"/encodeTmp/res
+	encodeImageAndCopy "res"  "encodeTmp/res"
+#cd proj.android
+	copyAssert encodeTmp/res assets
+	echo "scriptsVersion = $2" >$APP_ANDROID_ROOT/assets/config.lua
+	echo "module('release', package.seeall) release = true" > $APP_ANDROID_ROOT/assets/release.lua
+fi
+if [ $1 = 2 ]
+	then
+	copyAssert res assets
+	copyAssert unPackagedRes/music assets
+	copyAssert unPackagedRes/res assets
+#	//cd ..
+	rm -rf "$APP_ROOT"/encodeTmp
+	mkdir "$APP_ROOT"/encodeTmp
+	mkdir "$APP_ROOT"/encodeTmp/res
+	encodeImageAndCopy "res"  "encodeTmp/res"
+	encodeImageAndCopy "unPackagedRes/res"  "encodeTmp/res"
+#cd proj.android
+	copyAssert encodeTmp/res assets
+	echo "scriptsVersion = $2" >$APP_ANDROID_ROOT/assets/config.lua
+	echo "module('release', package.seeall) release = true ; isSelfServer =true;" > $APP_ANDROID_ROOT/assets/release.lua
+fi
 
-copyAssert res assets
-rm -rf "$APP_ROOT"/encodeTmp
-mkdir "$APP_ROOT"/encodeTmp
-mkdir "$APP_ROOT"/encodeTmp/res
-encodeImageAndCopy "res"  "encodeTmp/res"
-copyAssert encodeTmp/res assets
-
-copyAssert sound_android assets
-cp "$APP_ROOT"/sound/bgm01.mp3 assets
-cp "$APP_ROOT"/sound/bgm02.mp3 assets
-cp "$APP_ROOT"/sound/bgm03.mp3 assets
-
-echo "scriptsVersion = $2" >$APP_ANDROID_ROOT/assets/config.lua
-
-echo "appSrc = '$3'" >$APP_ANDROID_ROOT/assets/appChannel.lua
+if [ $1 = 3 ]
+	then
+	copyAssert res assets
+	copyAssert unPackagedRes/music assets
+	copyAssert unPackagedRes/res assets
+#	//cd ..
+	rm -rf "$APP_ROOT"/encodeTmp
+	mkdir "$APP_ROOT"/encodeTmp
+	mkdir "$APP_ROOT"/encodeTmp/res
+	encodeImageAndCopy "res"  "encodeTmp/res"
+	encodeImageAndCopy "unPackagedRes/res"  "encodeTmp/res"
+#cd proj.android
+	copyAssert encodeTmp/res assets
+	echo "scriptsVersion = $2" >$APP_ANDROID_ROOT/assets/config.lua
+	echo "module('release', package.seeall) release = true ;" > $APP_ANDROID_ROOT/assets/release.lua
+fi
